🚚 [projects] Move function `abortupdate` from `services` to `update` module
