+++ tests/unit/services/test_update.py
@@ -9,11 +9,11 @@ from cutty.filesystems.domain.path import Path as VirtualPath
 from cutty.projects.create import CreateProject
 from cutty.projects.create import LATEST_BRANCH
 from cutty.projects.create import UPDATE_BRANCH
+from cutty.projects.update import abortupdate
 from cutty.projects.update import continueupdate
 from cutty.projects.update import skipupdate
 from cutty.projects.update import updateproject
 from cutty.repositories.domain.repository import Repository as Template
-from cutty.services.update import abortupdate
 from cutty.util.git import Repository
 from tests.util.git import createbranches
 from tests.util.git import resolveconflicts
